Updated for 3.4 release
git-svn-id: https://svn.code.sf.net/p/openpfgw/code@371 368e28a5-49e5-40c5-be6c-d84989c25d7b

+++ pform/prmsieve/compress/compri7.c
@@ -41,6 +41,6 @@ int cp7expandstream(unsigned int*into, unsigned char const* buf, unsigned int ro
 	into = cp5expand(into, base, building); 
 	base +=30;
     }
-    return into-originto;
+    return (int) (into-originto);
 }
